+++ plugins/variants/src/MultiLinearVariantRenderer/MultiVariantRenderer.ts
@@ -4,7 +4,7 @@ import { featureSpanPx, renderToAbstractCanvas } from '@jbrowse/core/util'
 // locals
 import { getCol } from '../util'
 
-import type { Source } from '../util'
+import type { Source } from '../types'
 import type { RenderArgsDeserialized as FeatureRenderArgsDeserialized } from '@jbrowse/core/pluggableElementTypes/renderers/FeatureRendererType'
 import type { Feature } from '@jbrowse/core/util'
 import type { ThemeOptions } from '@mui/material'
