+++ plugins/variants/src/MultiVariantBaseRenderer.ts
@@ -1,7 +1,7 @@
 import FeatureRendererType from '@jbrowse/core/pluggableElementTypes/renderers/FeatureRendererType'
 import { renderToAbstractCanvas } from '@jbrowse/core/util'
 
-import type { Source } from './util'
+import type { Source } from './types'
 import type { RenderArgsDeserialized as FeatureRenderArgsDeserialized } from '@jbrowse/core/pluggableElementTypes/renderers/FeatureRendererType'
 import type { Feature } from '@jbrowse/core/util'
 import type { ThemeOptions } from '@mui/material'
