+++ plugins/variants/src/shared/Tooltip.tsx
@@ -6,7 +6,7 @@ import { observer } from 'mobx-react'
 import Tooltip from '../Tooltip'
 
 import type { TooltipContentsComponent } from '../Tooltip'
-import type { Source } from '../util'
+import type { Source } from '../types'
 import type { Feature } from '@jbrowse/core/util'
 
 const en = (n: number) => n.toLocaleString('en-US')
